requirements.txt
asvmq
pyserial
asvprotobuf
numpy
